+++ src/Element.js
@@ -1,35 +1,24 @@
 import styled from 'styled-components';
 
-import { count } from './perspective';
+import { layers } from './layers';
 
 export const Element = styled.div`
   ${({
     perspective = { x: 5, y: 5 },
     color = '#444',
-    hoverable = false,
+    hoverable = true,
     focusable = true
   }) => `
     transition: all 300ms ease;
 
-    &:hover {
-      cursor: pointer;
-    }
-
     &:not(:focus):not(:active) {
-      box-shadow: ${count(perspective.x, perspective.y, color)};
+      box-shadow: ${layers(perspective.x, perspective.y, color)};
     }
 
     ${hoverable &&
       `
-      &:not(:active):hover {
-        box-shadow: ${count(
-          perspective.x > 0 ? perspective.x + 2 : perspective.x - 2,
-          perspective.y > 0 ? perspective.y + 2 : perspective.y - 2,
-          color
-        )};
-        transform: translate(${perspective.x > 0 ? -2 : 2}px, ${
-        perspective.y > 0 ? -2 : 2
-      }px);
+      &:hover {
+        cursor: pointer;
       }
     `}
 
@@ -43,6 +32,17 @@ export const Element = styled.div`
     `}
   `}
 `;
+
+/*&:not(:active):hover {
+  box-shadow: ${layers(
+    perspective.x > 0 ? perspective.x + 2 : perspective.x - 2,
+    perspective.y > 0 ? perspective.y + 2 : perspective.y - 2,
+    color
+  )};
+  transform: translate(${perspective.x > 0 ? -2 : 2}px, ${
+  perspective.y > 0 ? -2 : 2
+}px);
+}*/
 
 /*{
   hoverable: 'up' || 'down' || 'reverseDir',