+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 1.0.0 alpha 1
 #### New Feature
-- added `perspective`, `color`, `hoverable`, `focusable` and `onScroll` props
+- added `perspective`, `color`, `hoverable` and `focusable` props