+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## 1.0.0 alpha 1
 #### New Feature
+- added 2 components: Text and Element
 - added `perspective`, `color`, `hoverable` and `focusable` props