+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Monument UI Changelog
 
-## 1.0.0 alpha 1
+## 0.1.0 alpha 1
 #### New Feature
 - added 2 components: Text and Element
-- added `perspective`, `color`, `hoverable` and `focusable` props
+- added `perspective`, `color`, `hoverable` and `clickable` props